Change return from long to int

+++ Utils/file_utils.h
@@ -6,7 +6,8 @@
 #include <string>
 #include <sys/stat.h>
 #include <sys/types.h>
+#include <unistd.h>
 
 const char* getFileSystemContents();
-long getFileSize(const char *fileName);
+int getFileSize(const char *fileName);
 #endif
